Re-add the required augeas lens
This file really lives in https://github.com/mwhahaha/augeas-pgpool
and is from commit 15cc096 in that repo.  It's copied here because the
upstream repo (https://github.com/mwhahaha/puppet-pgpool) had this file
as a submodule, and our librarian puppet couldn't retrieve it as a submodule
when building the modules/ directory.

Later versions of puppet-librarian seem to handle the :install_path key in
the Puppetfile.  If we were to switch, we can use that to put the augeas-pgpool
repo in the Puppetfile, and have it install its files under the files/ directory
of the puppet-pgpool repo.